Full Terraform. Pull data from s3.

+++ setup-tools/src/generator/main.cpp
@@ -14,7 +14,7 @@ constexpr size_t POLYNOMIAL_DEGREE = 0x10000;
 
 int main(int argc, char **argv)
 {
-    const uint polynomial_degree = argc > 1 ? strtol(argv[1], NULL, 0) : POLYNOMIAL_DEGREE;
+    const size_t polynomial_degree = argc > 1 ? strtol(argv[1], NULL, 0) : POLYNOMIAL_DEGREE;
     printf("initializing libff \n");
     libff::alt_bn128_pp::init_public_params();
     printf("calling compute generator poly \n");
